# services/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Báscula API
  version: "1.0.0"
  description: |
    Multi-tenant farm labour and payroll API. One binary, one Postgres, four
    roles. Isolation is row level security keyed on `app.farm_id`, set from the
    access token inside the transaction that serves the request.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/plots:
    get:
      tags: [plots]
      operationId: listPlots
      summary: List plots with their crops
      x-action: plots.read
      x-roles: [owner, admin, weigher]
      parameters:
        - $ref: "#/components/parameters/QueryQ"
        - $ref: "#/components/parameters/QueryStatus"
        - $ref: "#/components/parameters/QueryIncludeDeleted"
      responses:
        "200":
          description: The farm's plots.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Plot" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [plots]
      operationId: createPlot
      summary: Register a plot and what is planted in it
      description: |
        The plot and its crops arrive in one body because the form the owner
        fills in is one form: identity, location, what is planted, and the
        shape drawn on the map.

        A crop may name its type and variety by id or by name; a name the
        catalogue has not seen is added to it rather than rejected, which is
        the "add it if it is not there" button RSP-001 asks for.
      x-action: plots.write
      x-roles: [owner, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlotInput" }
      responses:
        "200":
          description: That id already existed here.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plot" }
        "201":
          description: Registered.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plot" }
        "400":
          description: |
            Malformed, or INVALID_GEOMETRY when the boundary is not a valid
            Polygon or MultiPolygon.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: DUPLICATE_NAME — this farm already has a plot with that name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/plots/{id}:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [plots]
      operationId: getPlot
      summary: One plot with its crops
      x-action: plots.read
      x-roles: [owner, admin, weigher]
      responses:
        "200":
          description: The plot.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plot" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [plots]
      operationId: updatePlot
      summary: Edit a plot, or take it out of service
      description: |
        A field that is absent keeps its value. Crops are not replaced from
        here — they have their own routes, so an edit to the plot's name
        cannot silently drop what is planted in it.
      x-action: plots.write
      x-roles: [owner, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlotPatch" }
      responses:
        "200":
          description: The plot as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plot" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            PLOT_HAS_ACTIVE_CROPS — a plot is not taken out of service while
            something is planted in it, because that would orphan the work
            records pointing at that crop. `details.activeCrops` says how many.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [plots]
      operationId: deletePlot
      summary: Take a plot out of service
      description: Logical. Equivalent to PATCH {"status":"inactive"}.
      x-action: plots.write
      x-roles: [owner, admin]
      responses:
        "204": { description: Marked inactive. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: PLOT_HAS_ACTIVE_CROPS.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/plots/{id}/boundary:
    parameters:
      - $ref: "#/components/parameters/PathID"
    put:
      tags: [plots]
      operationId: setPlotBoundary
      summary: Store the polygon drawn on the map
      description: |
        GeoJSON in, GeoJSON out. PostGIS stops at this boundary on purpose: the
        web and the phone never see a geography type, which is what keeps
        changing engines possible.

        Three PostGIS functions are used in this whole service. `ST_IsValid`
        rejects a ring that crosses itself before it can be stored, because a
        self-intersecting polygon has no area anybody would agree on and
        `computedAreaHa` would then be a confident lie. `ST_Area/10000`
        produces `computedAreaHa`, which comes back alongside the declared
        `areaHa` and never instead of it — the two always disagree, and
        choosing which to show is the owner's decision. `ST_Intersects` fills
        `overlaps`, which is a warning and never a refusal: two plots that
        touch are usually a drawing worth a second look, and sometimes they are
        a terrace above a coffee lot.

        A plain Polygon is promoted to a MultiPolygon: a plot in two separate
        pieces is ordinary.
      x-action: plots.boundary.write
      x-roles: [owner, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BoundaryInput" }
      responses:
        "200":
          description: Stored, with the recomputed hectares and any overlaps.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoundaryResult" }
        "400":
          description: |
            INVALID_GEOMETRY — not GeoJSON, not a polygon, or a ring that
            crosses itself. The message carries `ST_IsValidReason`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A 15-minute access token with claims `sub`, `farm_id`, `role`,
        `device_id`. The farm comes from here and from nowhere else.
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    QueryQ:
      name: q
      in: query
      description: Free-text needle, case-insensitive substring.
      schema: { type: string }
    QueryStatus:
      name: status
      in: query
      description: |
        `active` (the default), `inactive` for what has been logically removed,
        `all` for both.
      schema: { type: string, enum: [active, inactive, all] }
    QueryIncludeDeleted:
      name: includeDeleted
      in: query
      description: Legacy synonym for `status=all`, kept because the phone sends it.
      deprecated: true
      schema: { type: boolean }
...
  responses:
    BadRequest:
      description: BAD_REQUEST — malformed body, unknown field, or a bad value.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: UNAUTHORIZED — no token, or a token this server did not issue.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: FORBIDDEN — the permission table refused this role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        NOT_FOUND. Also what a resource of another farm answers, and what an id
        that is not a UUID answers: nothing with that id exists here.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ErrorCode:
      type: string
      description: |
        The machine-readable half of every failure. The client branches on
        this; the translation lives in the client, which is why `message` is
        English.

        This enum is compared against domain.AllCodes() by the contract test,
        so a code cannot exist in the server without appearing here.
      enum:
        - BAD_REQUEST
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL
        - TENANT_NOT_SET
        - INVALID_CREDENTIALS
        - EMAIL_NOT_VERIFIED
        - EMAIL_TAKEN
        - TOKEN_EXPIRED
        - TOKEN_REUSED
        - RATE_LIMITED
        - FARM_LIMIT_REACHED
        - FARM_SUSPENDED
        - WORK_RECORD_SETTLED
        - PAYABLE_ALREADY_CLAIMED
        - SETTLEMENT_ALREADY_VOID
        - ALREADY_REVERSED
        - NOTHING_TO_SETTLE
        - AMOUNT_EXCEEDS_BALANCE
        - INVALID_GEOMETRY
        - PLOT_HAS_ACTIVE_CROPS
        - NO_RATE_IN_FORCE
        - RANGE_NEEDS_FROZEN_RATE
        - DUPLICATE_DOCUMENT
        - DUPLICATE_NAME
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: English, for a developer and a log. Never shown to a picker.
            details:
              type: object
              additionalProperties: true
              description: |
                What a client needs to recover. The canonical case is
                PAYABLE_ALREADY_CLAIMED, where `winningSettlement` lets the
                phone re-derive its state.
...
    GeoJsonGeometry:
      type: object
      description: |
        A GeoJSON Polygon or MultiPolygon in WGS 84. PostGIS never crosses this
        boundary: the wire format is GeoJSON so the clients never see a
        geography type and changing engines stays possible.
      required: [type, coordinates]
      properties:
        type: { type: string, enum: [Polygon, MultiPolygon] }
        coordinates:
          type: array
          items: {}
      example:
        type: Polygon
        coordinates: [[[-75.88, 5.66], [-75.87, 5.66], [-75.87, 5.67], [-75.88, 5.67], [-75.88, 5.66]]]
    Plot:
      type: object
      required: [id, name, crops]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        areaHa:
          type: number
          format: double
          nullable: true
          description: What the owner declared.
        computedAreaHa:
          type: number
          format: double
          nullable: true
          description: |
            ST_Area of the polygon, in hectares. Returned alongside `areaHa`
            and never instead of it: they always disagree, and hiding one
            decides for the owner which one lies. Null until a boundary exists.
        department: { type: string, nullable: true }
        municipality: { type: string, nullable: true }
        boundary:
          allOf: [{ $ref: "#/components/schemas/GeoJsonGeometry" }]
          nullable: true
        createdAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
        crops:
          type: array
          items: { $ref: "#/components/schemas/PlotCrop" }
    PlotCrop:
      type: object
      required: [id, plotId, cropTypeId, cropType]
      properties:
        id: { type: string, format: uuid }
        plotId: { type: string, format: uuid }
        cropTypeId: { type: string, format: uuid }
        cropType: { type: string }
        varietyId: { type: string, format: uuid, nullable: true }
        variety: { type: string, nullable: true }
        areaHa: { type: number, format: double, nullable: true }
        plantedOn: { type: string, format: date-time, nullable: true }
        removedOn: { type: string, format: date-time, nullable: true }
        deletedAt: { type: string, format: date-time, nullable: true }
    PlotCropInput:
      type: object
      description: |
        Type and variety may be given by id or by name. A name the catalogue
        has not seen is added to it, which is the "add it if it is not there"
        button rather than a rejection.
      properties:
        id: { type: string, format: uuid }
        cropTypeId: { type: string, format: uuid }
        cropType: { type: string }
        varietyId: { type: string, format: uuid, nullable: true }
        variety: { type: string, nullable: true }
        areaHa: { type: number, format: double, nullable: true }
        plantedOn: { type: string, format: date-time, nullable: true }
        removedOn: { type: string, format: date-time, nullable: true }
    PlotInput:
      type: object
      required: [name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1 }
        areaHa: { type: number, format: double, nullable: true }
        department: { type: string, nullable: true }
        municipality: { type: string, nullable: true }
        boundary:
          allOf: [{ $ref: "#/components/schemas/GeoJsonGeometry" }]
          nullable: true
        crops:
          type: array
          items: { $ref: "#/components/schemas/PlotCropInput" }
    PlotPatch:
      allOf:
        - $ref: "#/components/schemas/PlotInput"
        - type: object
          properties:
            name: { type: string }
            status: { type: string, enum: [active, inactive] }
    BoundaryInput:
      type: object
      description: |
        `boundary` is the field; `geojson` is accepted as a synonym because the
        design document used that word.
      properties:
        boundary:
          allOf: [{ $ref: "#/components/schemas/GeoJsonGeometry" }]
        geojson:
          allOf: [{ $ref: "#/components/schemas/GeoJsonGeometry" }]
          deprecated: true
    BoundaryResult:
      type: object
      required: [plot, overlaps]
      properties:
        plot: { $ref: "#/components/schemas/Plot" }
        overlaps:
          type: array
          description: |
            Other plots this one now intersects. A warning, never a refusal.
          items: { $ref: "#/components/schemas/CatalogItem" }

    # -- catalogues and activities -------------------------------------------
    CatalogItem:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
...